web_dashboard/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==2.5.0
pydantic-settings>=2.1.0
pyjwt[crypto]==2.10.1
supabase==2.11.0
python-dotenv==1.0.0